+++ include/fused_kernel/core/utils/parameter_pack_utils.cuh
@@ -91,7 +91,7 @@ namespace fk { // namespace fused kernel
         return tuple_insert<sizeof...(Args) - 1>(t, Tuple<Args...>{args...});
     }
 
-    template <typename TransformType, typename SourceType, int NElems, int... Idx, typename... ExtraParams>
+    template <typename TransformType, typename SourceType, size_t NElems, int... Idx, typename... ExtraParams>
     constexpr inline std::array<typename TransformType::OutputType, NElems>
         static_transform_helper( const int& usedPlanes,
                                  const std::array<SourceType, NElems>& srcArray,
@@ -100,7 +100,7 @@ namespace fk { // namespace fused kernel
         return { (TransformType::template transform<Idx>(usedPlanes, srcArray[Idx], extParams...))... };
     }
 
-    template <typename TransformType, typename SourceType, int NElems, typename... ExtraParams>
+    template <typename TransformType, typename SourceType, size_t NElems, typename... ExtraParams>
     constexpr inline std::array<typename TransformType::OutputType, NElems>
         static_transform(const int& usedPlanes, const std::array<SourceType, NElems>& srcArray, const ExtraParams&... extParams) {
         return static_transform_helper<TransformType>(usedPlanes, srcArray, std::make_integer_sequence<int, NElems>{}, extParams...);
@@ -124,13 +124,13 @@ namespace fk { // namespace fused kernel
         }
     };
 
-    template <typename FT, typename ST, int NElems>
+    template <typename FT, typename ST, size_t NElems>
     constexpr inline std::array<FT, NElems>
         static_transform_get_first(const std::array<std::pair<FT, ST>, NElems>& srcArray) {
         return static_transform_helper<GetFirst<FT,ST>>(NElems, srcArray, std::make_integer_sequence<int, NElems>{});
     }
 
-    template <typename FT, typename ST, int NElems>
+    template <typename FT, typename ST, size_t NElems>
     constexpr inline std::array<ST, NElems>
         static_transform_get_second(const std::array<std::pair<FT, ST>, NElems>& srcArray) {
         return static_transform_helper<GetSecond<FT, ST>>(NElems, srcArray, std::make_integer_sequence<int, NElems>{});
